+++ lib/td_dq/rules/search_helper.ex
@@ -6,6 +6,7 @@ defmodule Search.Helpers do
   alias TdCache.ConceptCache
   alias TdCache.TaxonomyCache
   alias TdCache.UserCache
+  alias TdDq.Rules.Implementations
 
   def get_domain_ids(%{business_concept_id: nil}), do: -1
 
@@ -89,5 +90,15 @@ defmodule Search.Helpers do
 
   def with_result_text(result_map, _minimum, _goal, _type) do
     result_map
+  end
+
+  def get_sources(structure_ids) do
+    structure_ids
+    |> Enum.map(&Implementations.read_structure_from_cache/1)
+    |> Enum.filter(&(not Enum.empty?(&1)))
+    |> Enum.map(&Map.get(&1, :metadata, %{}))
+    |> Enum.map(&Map.get(&1, "alias"))
+    |> Enum.filter(& &1)
+    |> Enum.uniq()
   end
 end